fix(`__init__`): fixed an issue with weight matrix initialization that would lead to incorrect formatting of a layer's weight matrix
In essence, the issue was that the matrix dimensions were set up incorrectly, where the matrix was filled with lists of neurons per weight, instead of the other way around. Ideally, you'd want the weight matrix to consist of lists representing neurons and each list contain the weights of each neurons.