fix(`stochasticGD`): resolved an issue with incorrect computation of updated `self.biases`
### Explanation
The issue was that the usage of `numpy.einsum()` to reduce a `(n,b)` matrix into an `(n,1)` matrix actually converted the `(n,1)` into a 1-D array. Due to this unexpected transformation, NumPy improperly calculated the subtraction of `b` (which was an `(n,1)` matrix) and `gb` (which ended up being a 1-D array with a length of `n`). Since NumPy cannot compute the subtraction of a matrix and an array, it attempted to broadcast the array into a matrix which resulted in a matrix `gb` with the dimensions `(1,n)`. The subtraction in the end was a `self.biases` of `(n,n)` (individual unit) instead of the desired `(n,1)`.